# Deterministic patterns