If a pattern is fully deterministic and fully bound, we simply check that it's in the word list as opposed to the wasteful stuff we did before. We also improve the sort order.

Example: `cAcB;AB`. Note that the new version is faster and finds more solutions.

Before:
```
CHOCO • HOO
CIRCA • IRA
CIRCE • IRE
CISCO • ISO
CROCE • ROE
CROCI • ROI
CUSCO • USO
CUZCO • UZO
CHACHA • HAHA
COCKED • OKED
CRUTCH • RUTH
CANDICE • ANDIE
CATCALL • ATALL
CELICAS • ELIAS
CHANCES • HANES
CLICKED • LIKED
CONCEAL • ONEAL
CORACLE • ORALE
CRACKED • RAKED
CUTICLE • UTILE
Loaded 113096 words in 0.321s; solved in 0.637s (20 tuples).
```

After:
```
CHOCO • HOO
CIRCA • IRA
CIRCE • IRE
CISCO • ISO
CROCE • ROE
CROCI • ROI
CUSCO • USO
CUZCO • UZO
CHACHA • HAHA
COCKED • OKED
CRUTCH • RUTH
CANDICE • ANDIE
CATCALL • ATALL
CELICAS • ELIAS
CHANCES • HANES
CLICKED • LIKED
CONCEAL • ONEAL
CORACLE • ORALE
CRACKED • RAKED
CUTICLE • UTILE
CCCREAMS • CREAMS
CLICKING • LIKING
COCKSURE • OKSURE
CRESCENT • RESENT
CROCUSES • ROUSES
CRESCENTS • RESENTS
Loaded 113096 words in 0.295s; solved in 0.213s (26 tuples).
```